refactor: rename asset_plugin_gcr to amapy_plugin_gcr across the codebase
