+++ client/asset-core/asset_core/objects/tests/test_object_factory.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 from asset_core.objects.group.group_object import GroupObject
 from asset_core.objects.object import Object
 from asset_core.objects.object_factory import ObjectFactory
-from asset_plugin_gcr.gcr_storage import GcrStorage
+from amapy_plugin_gcr.gcr_storage import GcrStorage
 from asset_plugin_gcs.gcs_storage import GcsStorage
 from asset_plugin_s3.aws_storage import AwsStorage
