+++ client/asset-plugin-gcr/amapy_plugin_gcr/gcr_blob.py
@@ -1,5 +1,5 @@
 from amapy_pluggy.storage import StorageData
-from asset_plugin_gcr.gcr_url import GcrURL
+from amapy_plugin_gcr.gcr_url import GcrURL
 
 
 class GcrBlob(StorageData):
