+++ client/asset-plugin-gcr/amapy_plugin_gcr/gcr_storage.py
@@ -5,10 +5,10 @@ from amapy_pluggy.plugin import hook_impl
 from amapy_pluggy.plugin.object_content import ObjectContent
 from amapy_pluggy.storage import StorageData, StorageURL, Transporter
 from amapy_pluggy.storage.asset_storage import AssetStorage
-from asset_plugin_gcr.gcr_blob import GcrBlob
-from asset_plugin_gcr.gcr_mixin import GcrMixin
-from asset_plugin_gcr.gcr_url import GcrURL
-from asset_plugin_gcr.transporter.gcr_transporter import GcrTransporter
+from amapy_plugin_gcr.gcr_blob import GcrBlob
+from amapy_plugin_gcr.gcr_mixin import GcrMixin
+from amapy_plugin_gcr.gcr_url import GcrURL
+from amapy_plugin_gcr.transporter.gcr_transporter import GcrTransporter
 
 
 class GcrStorage(AssetStorage, GcrMixin):
