+++ client/asset-plugin-gcr/amapy_plugin_gcr/tests/test_gcr_storage.py
@@ -1,8 +1,8 @@
 import json
 from unittest.mock import patch
 
-from asset_plugin_gcr.gcr_blob import GcrBlob
-from asset_plugin_gcr.gcr_storage import GcrStorage
+from amapy_plugin_gcr.gcr_blob import GcrBlob
+from amapy_plugin_gcr.gcr_storage import GcrStorage
 
 
 def test_get_blob(fake_url_data):
@@ -21,7 +21,7 @@ def test_get_blob(fake_url_data):
         "url": "gcr.io/my-test-project/my-test-image@sha256:1234567890abcdef1234567890abcdef1234567890abcdef1234567890",
     }
 
-    with patch("asset_plugin_gcr.gcr_mixin.GcrMixin.fetch_url_data", return_value=fake_url_data):
+    with patch("amapy_plugin_gcr.gcr_mixin.GcrMixin.fetch_url_data", return_value=fake_url_data):
         for idx, url in enumerate(urls):
             blob: GcrBlob = GcrStorage.shared().get_blob(url=url)
             for key in expected:
@@ -46,7 +46,7 @@ def test_list_blobs(fake_url_data):
         "url": "gcr.io/my-test-project/my-test-image@sha256:1234567890abcdef1234567890abcdef1234567890abcdef1234567890",
     }
 
-    with patch("asset_plugin_gcr.gcr_mixin.GcrMixin.fetch_url_data", return_value=fake_url_data):
+    with patch("amapy_plugin_gcr.gcr_mixin.GcrMixin.fetch_url_data", return_value=fake_url_data):
         for idx, url in enumerate(urls):
             blob: GcrBlob = GcrStorage.shared().get_blob(url=url)
             for key in expected:
